PSGeneration_LLM/backend/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
python-docx==0.8.11
openai==1.1.0
numpy==1.25.2
Werkzeug==2.3.7
langchain==0.2.3
langchain-community==0.2.3
faiss-cpu==1.7.4
docx2python==2.1.6